+++ lib/core/utils/get_epoch_time.dart
@@ -1,0 +1,1 @@
+int get formattedEpochTime => DateTime.now().millisecondsSinceEpoch ~/ 1000;
